+++ app/src/index.tsx
@@ -83,6 +83,7 @@ function UuiEnhancedApp() {
         router,
         apiReloginPath: 'api/auth/login',
         apiPingPath: 'api/auth/ping',
+        shadowRootHost: document.getElementById('root')?.shadowRoot,
     });
 
     useEffect(() => {